Updates for 60k accessions

+++ modules/assembly.nf
@@ -1,9 +1,18 @@
-process spades_single_end {
-    publishDir 'single_end', mode: 'copy'
+process SPADES_SINGLE_END {
+    tag "${bioproj_id}"
+    publishDir "${params.publish_dir}/single_end"
+
     container 'docker://staphb/spades'
     conda 'bioconda::spades=3.15.5'
 
-    memory '32 GB'
+    cpus 16
+    memory { 16.GB * task.attempt }
+    time { task.attempt > 1 ? 24.hour : 6.hour }
+    maxRetries 5
+    errorStrategy {
+        (task.exitStatus == 9 || task.attempt > maxRetries) ? 'ignore' : 'retry' 
+    }
+    scratch true
 
     input:
     tuple val(bioproj_id), path(fq)
@@ -14,16 +23,31 @@ process spades_single_end {
     script:
     """
     spades.py --rna -s $fq -o .
-    mv transcripts.fasta ${bioproj_id}_S.transcripts.fasta
+    if [ -f transcripts.fasta ]; then
+        sed -i 's/>NODE/>${bioproj_id}_S_NODE/g' transcripts.fasta
+        mv transcripts.fasta ${bioproj_id}_S.transcripts.fasta
+    else
+        echo 'No transcripts.fasta file found!'
+        exit 9
+    fi
     """
 }
 
-process spades_paired_end {
-    publishDir 'paired_end', mode: 'copy'
+process SPADES_PAIRED_END {
+    tag "${bioproj_id}"
+    publishDir "${params.publish_dir}/paired_end"
+
     container 'docker://staphb/spades'
     conda 'bioconda::spades=3.15.5'
 
-    memory '32 GB'
+    cpus 16
+    memory { 32.GB * task.attempt }
+    time { task.attempt > 1 ? 24.hour : 6.hour }
+    maxRetries 5
+    errorStrategy {
+        (task.exitStatus == 9 || task.attempt > maxRetries) ? 'ignore' : 'retry' 
+    }
+    scratch true
 
     input:
     tuple val(bioproj_id), path(fq1), path(fq2)
@@ -34,6 +58,12 @@ process spades_paired_end {
     script:
     """
     spades.py --rna -1 $fq1 -2 $fq2 -o .
-    mv transcripts.fasta ${bioproj_id}_P.transcripts.fasta
+    if [ -f transcripts.fasta ]; then
+        sed -i 's/>NODE/>${bioproj_id}_P_NODE/g' transcripts.fasta
+        mv transcripts.fasta ${bioproj_id}_P.transcripts.fasta
+    else
+        echo 'No transcripts.fasta file found!'
+        exit 9
+    fi
     """
 }
